+++ browser/components/SimpleContributions/steps/TicketSubmission.tsx
@@ -11,8 +11,8 @@ export default class TicketSubmission extends SurveyStep<{}> {
         <ExtensionPoint ext="simple-contrib-auto-approve-deny">
           <p>
             Your change was not able to be automatically approved. You must
-            follow the regular <Link to="/contribute">Log Contribution</Link>
-            process.
+            follow the regular <Link to="/contribute">Log Contribution </Link>
+             process.
           </p>
         </ExtensionPoint>
       </div>