+++ browser/config.js
@@ -1,0 +1,9 @@
+export const browser = {
+  autoApprove: {
+    // contribution types allowed for simple contribution
+    // allowed options: bugfix, feature, doc, config, test, other
+    allowedTypes: ['bugfix', 'doc', 'config', 'test'],
+    // contribDescTypes: used in ContribTypes for steps that require additional contributions details
+    contribDescTypes: ['feature', 'other'],
+  }
+};